TaskScheduler class from the services/init.py file into utils/task_scheduler.py
